uart: split between esp32 and esp8266.
