+++ esphome/components/uart/uart.h
@@ -13,6 +13,8 @@ enum UARTParityOptions {
   UART_CONFIG_PARITY_ODD,
 };
 
+const char *parity_to_str(UARTParityOptions parity);
+
 #ifdef ARDUINO_ARCH_ESP8266
 class ESP8266SoftwareSerial {
  public:
